First working implementation of the MeshAsset Spec!

+++ Graphics/Include/DenOfIzGraphics/Assets/Stream/BinaryWriter.h
@@ -49,10 +49,10 @@ namespace DenOfIz
         DZ_API void                   WriteBytes( const InteropArray<Byte> &buffer ) const;
         DZ_API void                   WriteUInt16( uint16_t value ) const;
         DZ_API void                   WriteUInt32( uint32_t value ) const;
-        DZ_API void                   WriteUInt64( uint32_t value ) const;
+        DZ_API void                   WriteUInt64( uint64_t value ) const;
         DZ_API void                   WriteInt16( int16_t value ) const;
         DZ_API void                   WriteInt32( int32_t value ) const;
-        DZ_API void                   WriteInt64( int32_t value ) const;
+        DZ_API void                   WriteInt64( int64_t value ) const;
         DZ_API void                   WriteFloat( float value ) const;
         DZ_API void                   WriteString( const InteropString &value ) const;
         DZ_API void                   WriteUInt16_2( const UInt16_2 &value ) const;
